# public/api-docs.yaml
openapi: 3.0.0
info:
  title: TheoBot Database API Layer
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [TelegramBot API Docs](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
servers:
  - url: https://dqpllfj7tb.execute-api.us-east-1.amazonaws.com/dev
paths:
...
  /moons/{id}:
    get:
      tags:
        - Moons
      summary: Get a moon by ID
      operationId: getMoon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about a moon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Moon'
        '404':
          description: Moon not found
    put:
      tags:
        - Moons
      summary: Update a moon
      operationId: updateMoon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Moon'
      responses:
        '200':
          description: Moon updated
        '400':
          description: Invalid input
        '404':
          description: Moon not found
    delete:
      tags:
        - Moons
      summary: Delete a moon
      operationId: deleteMoon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Moon deleted
        '404':
          description: Moon not found
  /moons/byPlanet/{planetID}:
    get:
      tags:
        - Moons
      summary: Get moons by planet ID
      operationId: moonsByPlanetID
      parameters:
        - name: planetID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: An array of moons associated with a planet ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Moon'
  /planets:
    get:
      tags:
        - Planets
      summary: List all planets
      operationId: listPlanets
      responses:
        '200':
          description: An array of planets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Planet'
    post:
      tags:
        - Planets
      summary: Create a new planet
      operationId: createPlanet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Planet'
      responses:
        '201':
          description: Planet created
        '400':
          description: Invalid input
  /planets/{id}:
    get:
      tags:
        - Planets
      summary: Get a planet by ID
      operationId: getPlanet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about a planet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Planet'
        '404':
          description: Planet not found
    put:
      tags:
        - Planets
      summary: Update a planet
      operationId: updatePlanet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Planet'
      responses:
        '200':
          description: Planet updated
        '400':
          description: Invalid input
        '404':
          description: Planet not found
    delete:
      tags:
        - Planets
      summary: Delete a planet
      operationId: deletePlanet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Planet deleted
        '404':
          description: Planet not found
  /files:
    get:
      tags:
        - Files
      summary: List all files
      operationId: listFiles
      responses:
        '200':
          description: An array of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/File'
    post:
      tags:
        - Files
      summary: Create a new file
      operationId: createFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/File'
      responses:
        '201':
          description: File created
        '400':
          description: Invalid input
  /files/{id}:
    get:
      tags:
        - Files
      summary: Get a file by ID
      operationId: getFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about a file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          description: File not found
    put:
      tags:
        - Files
      summary: Update a file
      operationId: updateFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/File'
      responses:
        '200':
          description: File updated
        '400':
          description: Invalid input
        '404':
          description: File not found
    delete:
      tags:
        - Files
      summary: Delete a file
      operationId: deleteFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File deleted
        '404':
          description: File not found
  /files/moon/{moonId}:
    get:
      tags:
        - Files
      summary: List all files associated with a specific moon
      parameters:
        - name: moonId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of files
        '500':
          description: Error retrieving moon files
  /files/task/{taskId}:
    get:
      tags:
        - Files
      summary: List all files associated with a specific task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of files
        '500':
          description: Error retrieving task files
  /files/project/{projectId}:
    get:
      tags:
        - Files
      summary: List all files associated with a specific project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of files
        '500':
          description: Error retrieving project files
...
  /chats:
    get:
      tags:
        - Chats
      summary: List all chat messages
      operationId: listChats
      responses:
        '200':
          description: An array of chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chat'
    post:
      tags:
        - Chats
      summary: Create a new chat message
      operationId: createChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        '201':
          description: Chat message created
        '400':
          description: Invalid input
  /chats/{id}:
    get:
      tags:
        - Chats
      summary: Get a chat message by ID
      operationId: getChat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about a chat message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '404':
          description: Chat message not found
    put:
      tags:
        - Chats
      summary: Update a chat message
      operationId: updateChat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        '200':
          description: Chat message updated
        '400':
          description: Invalid input
        '404':
          description: Chat message not found
    delete:
      tags:
        - Chats
      summary: Delete a chat message
      operationId: deleteChat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chat message deleted
        '404':
          description: Chat message not found
  /events:
    get:
      tags:
        - Events
      summary: List all events
      operationId: listEvents
      responses:
        '200':
          description: An array of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      tags:
        - Events
      summary: Create a new event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Event created
        '400':
          description: Invalid input
  /events/{id}:
    get:
      tags:
        - Events
      summary: Get an event by ID
      operationId: getEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about an event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    put:
      tags:
        - Events
      summary: Update an event
      operationId: updateEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: Event updated
        '400':
          description: Invalid input
        '404':
          description: Event not found
    delete:
      tags:
        - Events
      summary: Delete an event
      operationId: deleteEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event deleted
        '404':
          description: Event not found
components:
  schemas:
    Satellite:
      type: object
      required:
        - id
        - moonID
        - title
      properties:
        id:
          type: string
          description: Unique identifier for the Satellite
        moonID:
          type: string
          description: Identifier for the Moon that the Satellite is associated with
        title:
          type: string
          description: Title of the Satellite
        description:
          type: string
          description: Description of the Satellite
        User:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required:
        - id
        - firstName
        - username
        - completedOnboarding
        - role
        - status
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        completedOnboarding:
          type: boolean
        role:
          $ref: '#/components/schemas/UserRole'
        bio:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
    UserRole:
      type: string
      enum:
        - SUPER_ADMIN
        - ADMIN
        - MEMBER
        - VIEWER
    UserStatus:
      type: string
      enum:
        - PENDING
        - CONFIRMED
        - SUSPENDED
    Moon:
      type: object
      required:
        - id
        - name
        - planetID
      properties:
        id:
          type: string
        name:
          type: string
        planetID:
          type: string
        satellites:
          type: array
          items:
            $ref: '#/components/schemas/Satellite'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        chats:
          type: array
          items:
            $ref: '#/components/schemas/Chat'
        eventID:
          type: string
    Project:
      type: object
      required:
        - id
        - moonID
        - name
        - status
        - projectLead
      properties:
        id:
          type: string
        moonID:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - COMPLETED
            - ON_HOLD
        projectLead:
          type: string
          description: Name of the project lead. Used for Jira project creation.
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRole'
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        chats:
          type: array
          items:
            $ref: '#/components/schemas/Chat'
    Planet:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        chats:
          type: array
          items:
            $ref: '#/components/schemas/Chat'
        moons:
          type: array
          items:
            $ref: '#/components/schemas/Moon'
        eventID:
          type: string
    File:
      type: object
      required:
        - id
        - name
        - url
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        uploadedBy:
          $ref: '#/components/schemas/User'
    Task:
      type: object
      required:
        - id
        - projectID
        - title
        - status
      properties:
        id:
          type: string
        projectID:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        assignedUsers:
          type: array
          items:
            $ref: '#/components/schemas/User'
        dueDate:
          type: string
          format: date
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        chats:
          type: array
          items:
            $ref: '#/components/schemas/Chat'
        notes:
          type: array
          items:
            type: string
    TaskStatus:
      type: string
      enum:
        - TO_DO
        - IN_PROGRESS
        - DONE
    Chat:
      type: object
      required:
        - id
        - message
        - sentAt
        - sentBy
      properties:
        id:
          type: string
        message:
          type: string
        sentAt:
          type: string
          format: date-time
        sentBy:
          $ref: '#/components/schemas/User'
    Event:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        date:
          type: array
          items:
            type: string
            format: date
        time:
          type: array
          items:
            type: string
        location:
          type: string
        attendees:
          type: array
          items:
            $ref: '#/components/schemas/User'
        moons:
          type: array
          items:
            $ref: '#/components/schemas/Moon'
        planets:
          type: array
          items:
            $ref: '#/components/schemas/Planet'
    ProjectRole:
      type: object
      required:
        - id
        - projectID
        - title
      properties:
        id:
          type: string
        projectID:
          type: string
          description: The ID of the project this role is associated with.
        title:
          type: string
          description: The title of the project role.
        description:
          type: string
          description: A brief description of the role.
        assignedTo:
          $ref: '#/components/schemas/User'
          description: The user who is assigned to this role.